Add files via upload
I have various js files as leaving all my array examples in the same file would have resulted in the const 'numbers' not being able to work which is crucial in an array function. Also, the file 'one.js' includes pass task 1 as well as my first array of task 2.